Implementing the Dashboard.js
The Dashboard accomodates all the coponents for the moment. A few components have been imported from the '@material-ui/core'. Also, the makeStyles which allows for css in js.
No functionality has been implemented yet, only the design of the chat application to give an idea on how it will look and operate.